requirements.txt
langchain==0.0.188
openai==0.27.7
google-api-python-client==2.88.0
python-dotenv==1.0.0
flask==2.3.2